ref(issues): Move multiple endpoints and tests to issues folder
- Move group_attachments, team_all_unresolved_issues, team_issue_breakdown, organization_event_details, organization_issues_resolved_in_release endpoints to issues/endpoints/
- Move issue_search module from api/ to issues/
- Move corresponding test files to tests/sentry/issues/endpoints/
- Update all import references in api/urls.py and other affected files

refs https://linear.app/getsentry/issue/RTC-1110/move-issues-endpoints-into-issues-app
