+++ src/sentry/issues/endpoints/group_attachments.py
