+++ src/sentry/workflow_engine/endpoints/organization_detector_index.py
@@ -19,7 +19,6 @@ from sentry.api.bases import OrganizationAlertRulePermission, OrganizationEndpoi
 from sentry.api.event_search import SearchConfig, SearchFilter, SearchKey, default_config
 from sentry.api.event_search import parse_search_query as base_parse_search_query
 from sentry.api.exceptions import ResourceDoesNotExist
-from sentry.api.issue_search import convert_actor_or_none_value
 from sentry.api.paginator import OffsetPaginator
 from sentry.api.serializers import serialize
 from sentry.apidocs.constants import (
@@ -35,6 +34,7 @@ from sentry.constants import ObjectStatus
 from sentry.deletions.models.scheduleddeletion import RegionScheduledDeletion
 from sentry.incidents.grouptype import MetricIssue
 from sentry.issues import grouptype
+from sentry.issues.issue_search import convert_actor_or_none_value
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.models.team import Team
