+++ tests/sentry/issues/endpoints/test_group_attachments.py
